+++ lib/src/features/initialization/application/app_startup.dart
@@ -2,7 +2,7 @@ import 'dart:io' as io;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_boolean_template/src/features/initialization/service/info_service.dart';
+import 'package:flutter_boolean_template/src/features/initialization/application/info_service.dart';
 import 'package:flutter_boolean_template/src/features/settings/data/dto/navigation_type_override.dart';
 import 'package:flutter_boolean_template/src/features/settings/data/dto/settings.dart';
 import 'package:flutter_boolean_template/src/features/settings/data/repository/settings_repository.dart';
