+++ lib/src/features/initialization/presentation/app_startup_widget.dart
@@ -1,6 +1,6 @@
 import 'package:constants/constants.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_boolean_template/src/features/initialization/service/app_startup.dart';
+import 'package:flutter_boolean_template/src/features/initialization/application/app_startup.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class AppStartupWidget extends ConsumerWidget {
