+++ lib/src/features/settings/presentation/widgets/settings_widget.dart
@@ -2,7 +2,7 @@ import 'package:constants/constants.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_boolean_template/src/features/initialization/service/info_service.dart';
+import 'package:flutter_boolean_template/src/features/initialization/application/info_service.dart';
 import 'package:flutter_boolean_template/src/features/settings/application/settings_service.dart';
 import 'package:flutter_boolean_template/src/features/settings/data/dto/navigation_type_override.dart';
 import 'package:flutter_boolean_template/src/features/settings/data/dto/settings.dart';
